Improve up next queue handling on episode screen
